% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{east_most_tile_x}
\alias{east_most_tile_x}
\title{Get the x index of the east most tile for a zoom level}
\usage{
east_most_tile_x(zoom)
}
\arguments{
\item{zoom}{an integer in the range 1:18}
}
\value{

}
\description{
Get the x index of the east most tile for a zoom level
}
