% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_data.R
\name{label_data}
\alias{label_data}
\title{Label data using a named list}
\usage{
label_data(df, labels)
}
\arguments{
\item{df}{data to label}

\item{labels}{a named list}
}
\value{
a dataframe
}
\description{
Label data using a named list
}
\examples{
df <- data.frame(a = c(1,2,3), b=c(4,5,6), c=c(7,8,9))
labels <- list(a = "col 1", b = "col 2", c = "col 3")
new_df <- label_data(df, labels)
}
