% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lon_wrap_180}
\alias{lon_wrap_180}
\title{Wrap coordinates that are on the anti-meridian (date line)}
\usage{
lon_wrap_180(sf)
}
\arguments{
\item{sf}{object}
}
\value{
a sf object with the same  projection as supplied
}
\description{
features in tiles that are the east most for a zoom level
(x is 2^(zoom) - 1) or the west most (x is 0) have longitudes
that go from positive to negative (e.g 179 to -182). This function
converts to a projection (+proj=longlat +datum=WGS84 +lon_wrap=180) that
wraps these longitudes and then converts back to the original projection
}
