% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_grid.R
\name{tile_grid}
\alias{tile_grid}
\title{Generate a grid of nine tiles the top left of centre tile is
the supplied longitude and latitude}
\usage{
tile_grid(bbox, zoom)
}
\arguments{
\item{bbox}{the bounding box to fit onto a grid of tiles. Must be either a
'bbox' object created with sf::st_bbox or a vector of length 4 with names:
'xmin', 'xmax', 'ymin', 'ymax'.}

\item{zoom}{integer required zoom level 1- 17}
}
\value{
a list of simple features
}
\description{
Generate a grid of nine tiles the top left of centre tile is
the supplied longitude and latitude
}
