% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victor-package.R
\docType{package}
\name{victor-package}
\alias{victor}
\alias{victor-package}
\title{victor: Work With Mapbox Vector Tiles In R}
\description{
Import data from the Mapbox Vector Tile API and convert
  to simple features.  Plot static maps using layers you find useful.
}
\author{
\strong{Maintainer}: Adam Gruer \email{adamgruer@icloud.com}

}
\keyword{internal}
