% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VICTOR.R
\name{VICTOR}
\alias{VICTOR}
\title{This function allows you to diagnose cell annotation labels.}
\usage{
VICTOR(
  seuratObj_Query,
  seuratObj_Ref,
  ActualCellTypeColumn = "Actual_Cell_Type",
  AnnotCellTypeColumn = "Annotation",
  Add_nROC = FALSE,
  seurat_version = "V5",
  ...
)
}
\arguments{
\item{seuratObj_Query}{Seurat Object of Query.}

\item{seuratObj_Ref}{Seurat Object of Reference.}

\item{ActualCellTypeColumn}{Column name of the actual cell type in meta.data.}

\item{AnnotCellTypeColumn}{Column name of the cell type annotation in meta.data.}
}
\description{
This function allows you to diagnose cell annotation labels.
}
\examples{
VICTOR(seuratObj_Query, seuratObj_Ref, ActualCellTypeColumn = "Actual_Cell_Type", AnnotCellTypeColumn = "Annotation", Add_nROC = FALSE)
}
\keyword{scRNA-seq}
